// mobile/src/api/generated/shop/schemas/patchedProfileUpdate.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * ShopDjango API
 * API for ShopDjango project
 * OpenAPI spec version: 1.0.0
 */
import type { RoleEnum } from './roleEnum';

export interface PatchedProfileUpdate {
  /**
   * User's first name
   * @maxLength 150
   */
  first_name?: string;
  /**
   * User's last name
   * @maxLength 150
   */
  last_name?: string;
  /**
   * User's date of birth
   * @nullable
   */
  date_of_birth?: string | null;
  /**
   * Phone number in international format. Supports digits, spaces, hyphens, parentheses, and optional '+' prefix. Example: +48131012012
   * @pattern ^\+?[\d\s\-\(\)]+$
   */
  phone_number?: string;
  /**
   * @minimum -2147483648
   * @maximum 2147483647
   */
  role?: RoleEnum;
}